feat: validação de erros, controller, dto, service de product

+++ dscommerce/src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseEXception;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,7 +25,9 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
 
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository
+                .findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado!"));
 
         return new ProductDTO(product);
     }
@@ -28,7 +35,8 @@ public class ProductService {
     @Transactional(readOnly = true)
     public Page<ProductDTO> findAll(Pageable pageable){
 
-        Page<Product> result = productRepository.findAll(pageable);
+        Page<Product> result = productRepository
+                .findAll(pageable);
 
         return result.map(x -> new ProductDTO(x));
     }
@@ -45,12 +53,15 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO){
+        try{
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, product);
+            product = productRepository.save(product);
 
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, product);
-        product = productRepository.save(product);
-
-        return new ProductDTO(product);
+            return new ProductDTO(product);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product product) {
@@ -60,9 +71,17 @@ public class ProductService {
         product.setImgUrl(productDTO.getImgUrl());
     }
 
-    @Transactional()
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
 
-        productRepository.deleteById(id);
+        if(!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+
+        try{
+            productRepository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseEXception("Falha na integridade referencial.");
+        }
     }
 }
